requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
